guru: document display page, ui improvements.

+++ docs/generate_docs.sh
@@ -1,11 +1,15 @@
 #!/bin/bash
 
+FULL=0
+
 set -x
 
 DOCS=(es1.1 es2.0 es3 es3.0 es3.1 gl2.1 gl4)
 
-rm -rf build
-mkdir build
+if [ "$FULL" -gt 0 ]; then
+    rm -rf build
+fi
+mkdir -p build
 cd build
 
 mkdir assets
@@ -14,10 +18,14 @@ do
     mkdir "assets/${doc}"
 done
 
-git clone --depth 1 https://github.com/KhronosGroup/OpenGL-Refpages.git
+if [ ! -e OpenGL-Refpages ]; then
+    git clone --depth 1 https://github.com/KhronosGroup/OpenGL-Refpages.git
+fi
 cd OpenGL-Refpages
 
 make || true
+
+set -e
 
 cp es1.1/xhtml/*.xml ../assets/es1.1/
 cp es2.0/xhtml/*.xml ../assets/es2.0/
@@ -37,4 +45,5 @@ python clean_docs.py build/assets/es3.1/
 python clean_docs.py build/assets/gl2.1/
 python clean_docs.py build/assets/gl4/
 
-python generate_index.py build/assets/ > build/assets/index.json
+rm -rf build/assets/index.json
+python generate_index.py build/assets/ -o build/assets/index.json
